requirements.txt
streamlit
sqlalchemy
openai
langchain
python-dotenv
boto3
pydub
ffmpeg


streamlit-authenticator
Werkzeug

